+++ src/efusor/fusor.py
@@ -14,31 +14,6 @@ from efusor.borda import borda
 from efusor.priority import prioritize
 
 
-def vectorize(labels: list[str], scores: dict | list) -> np.ndarray:
-    """
-    convert sequences of scores & labels into a vector (1d numpy array)
-    :param labels: ordered list of labels/classes for vectorization
-    :type labels: list[str]
-    :param scores: predictions as {label: score}
-    :type scores: dict | list
-    :return: array
-    :rtype: np.ndarray
-    """
-    return (batch(*[vectorize(labels, x) for x in scores]) if isinstance(scores, list) else
-            np.array([scores.get(x, np.nan) for x in labels]))
-
-
-def batch(*vector: np.ndarray) -> np.ndarray:
-    """
-    batch vectors into a matrix
-    :param vector:
-    :type vector: np.ndarray
-    :return: matrix
-    :rtype: np.ndarray
-    """
-    return np.stack(vector)
-
-
 def fuse(tensor: list | np.ndarray,
          method: str = "hard_voting",
          weights: list | np.ndarray = None
